requirements.txt
psycopg2-binary>=2.8
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.7
djangorestframework==3.12.4
httpie==2.5.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
pyaml==21.8.3
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
